Refactored instrumention tests to Match modifications
