+++ project/app/src/main/res/layout/row_match_info.xml
@@ -3,7 +3,6 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/matchActivity"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/second_color">
