+++ project/app/src/main/res/layout/row_match_sets.xml
@@ -2,7 +2,6 @@
 <android.support.constraint.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/matchActivity"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/second_color">
